moved the components inside the financial product folder
